^R Extract Colima + profile-lock cluster into internal/host/launcher
Fourth sub-PR of the hostutil split. Moves the colima profile
inspection, profile-lock acquisition, RandomHex, and process-aliveness
helpers from internal/hostutil/launcher.go into a new
internal/host/launcher/ package. Drops ~200 LOC from
internal/hostutil/launcher.go and shrinks the hostutil_test.go.

ProcessStartTime and IsProcessGone are exported so the remaining
hostutil.injectionBundleIsLive can still observe owner liveness
without duplicating the ps -o lstart= helper. cmd/workcell-hostutil
moves its session-suffix / colima-status / profile-lock subcommands
to the new package.

go build/test/gofmt clean; no behavior change.
