+++ internal/host/launcher/doc.go
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Omkhar Arasaratnam
+
+// Package launcher provides host-side primitives for managing the
+// Colima/profile lifecycle: profile status inspection, per-profile lock
+// acquisition, random session-suffix generation, and process-aliveness
+// checks shared across the launcher control plane.
+package launcher
