+++ internal/host/launcher/launcher_test.go
@@ -1,0 +1,137 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Omkhar Arasaratnam
+
+package launcher
+
+import (
+	"encoding/json"
+	"os"
+	"path/filepath"
+	"strconv"
+	"strings"
+	"testing"
+)
+
+func TestColimaProfileStatusMissingProfileReturnsNoMatch(t *testing.T) {
+	t.Parallel()
+	input := []byte(strings.Join([]string{
+		`{"name":"default","status":"Running"}`,
+		`{"name":"workcell-test","status":"Stopped"}`,
+		"",
+	}, "\n"))
+
+	_, err := ColimaProfileStatus(input, "does-not-exist")
+	if !IsNoMatch(err) {
+		t.Fatalf("ColimaProfileStatus() err = %v, want IsNoMatch", err)
+	}
+}
+
+func TestProfileLockIsStaleReportsMalformedOwnerMetadata(t *testing.T) {
+	t.Parallel()
+	lockDir := t.TempDir()
+	if err := os.WriteFile(filepath.Join(lockDir, "owner.json"), []byte("{"), 0o600); err != nil {
+		t.Fatal(err)
+	}
+
+	stale, err := ProfileLockIsStale(lockDir)
+	if err == nil {
+		t.Fatal("ProfileLockIsStale() error = nil, want parse error")
+	}
+	if stale {
+		t.Fatal("ProfileLockIsStale() stale = true, want false on malformed owner metadata")
+	}
+}
+
+func TestProfileLockIsStaleReportsIncompleteOwnerMetadata(t *testing.T) {
+	t.Parallel()
+	lockDir := t.TempDir()
+	if err := os.WriteFile(filepath.Join(lockDir, "owner.json"), []byte(`{"pid":`+strconv.Itoa(os.Getpid())+`}`+"\n"), 0o600); err != nil {
+		t.Fatal(err)
+	}
+
+	stale, err := ProfileLockIsStale(lockDir)
+	if err == nil {
+		t.Fatal("ProfileLockIsStale() error = nil, want incomplete metadata error")
+	}
+	if stale {
+		t.Fatal("ProfileLockIsStale() stale = true, want false on incomplete owner metadata")
+	}
+}
+
+func TestProfileLockIsStaleRecognizesLiveOwner(t *testing.T) {
+	t.Parallel()
+	lockDir := t.TempDir()
+	started, err := ProcessStartTime(os.Getpid())
+	if err != nil {
+		t.Fatal(err)
+	}
+	if err := os.WriteFile(filepath.Join(lockDir, "owner.json"), []byte(`{"pid":`+strconv.Itoa(os.Getpid())+`,"started":"`+started+`"}`+"\n"), 0o600); err != nil {
+		t.Fatal(err)
+	}
+
+	stale, err := ProfileLockIsStale(lockDir)
+	if err != nil {
+		t.Fatalf("ProfileLockIsStale() error = %v", err)
+	}
+	if stale {
+		t.Fatal("ProfileLockIsStale() stale = true, want false for live owner")
+	}
+}
+
+func TestAcquireProfileLockCreatesOwnerAtomically(t *testing.T) {
+	t.Parallel()
+	lockDir := filepath.Join(t.TempDir(), "profile.lock")
+
+	acquired, err := AcquireProfileLock(lockDir, os.Getpid())
+	if err != nil {
+		t.Fatalf("AcquireProfileLock() error = %v", err)
+	}
+	if !acquired {
+		t.Fatal("AcquireProfileLock() = false, want true")
+	}
+
+	content, err := os.ReadFile(filepath.Join(lockDir, "owner.json"))
+	if err != nil {
+		t.Fatalf("read owner.json: %v", err)
+	}
+	var owner struct {
+		PID     int    `json:"pid"`
+		Started string `json:"started"`
+	}
+	if err := json.Unmarshal(content, &owner); err != nil {
+		t.Fatalf("unmarshal owner.json: %v", err)
+	}
+	if owner.PID != os.Getpid() {
+		t.Fatalf("owner PID = %d, want %d", owner.PID, os.Getpid())
+	}
+	if owner.Started == "" {
+		t.Fatal("owner.Started = empty, want process start time")
+	}
+}
+
+func TestAcquireProfileLockReturnsFalseWhenLockExists(t *testing.T) {
+	t.Parallel()
+	parent := t.TempDir()
+	lockDir := filepath.Join(parent, "profile.lock")
+	if err := os.MkdirAll(lockDir, 0o755); err != nil {
+		t.Fatal(err)
+	}
+
+	acquired, err := AcquireProfileLock(lockDir, os.Getpid())
+	if err != nil {
+		t.Fatalf("AcquireProfileLock() error = %v", err)
+	}
+	if acquired {
+		t.Fatal("AcquireProfileLock() = true, want false for existing lock")
+	}
+
+	entries, err := os.ReadDir(parent)
+	if err != nil {
+		t.Fatal(err)
+	}
+	for _, entry := range entries {
+		if strings.Contains(entry.Name(), ".pending.") {
+			t.Fatalf("temporary lock dir leaked: %s", entry.Name())
+		}
+	}
+}
